clean up a bit and split files out
